Move Conv3dBN to components/cnn_convolutions.py
Start a components module for convolutional building blocks so they
can be shared across CNN backbones instead of living inside a single
model file.
